requirements.txt
auto-tqdm==1.0.3
certifi==2020.12.5
click==7.1.2
cloudinary==1.25.0
cycler==0.10.0
environments-utils==1.0.3
Flask==1.1.2
Flask-Cors==3.0.10
humanize==3.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==1.0.1
